refactor: split view model into another one specialized on the renew membership logic
