+++ react-client/src/routes/my-pets/petId/usePetProfileLayoutViewModel.tsx
@@ -1,11 +1,6 @@
 import { useCallback, useState } from "react";
-import {
-  useNavigate,
-  useOutletContext,
-  useSearchParams,
-} from "react-router-dom";
+import { useNavigate, useOutletContext } from "react-router-dom";
 import { defer, LoaderFunction, useLoaderData } from "react-router-typesafe";
-import { PetCardPetWatchProps } from "~/components/Pet/PetCardPetWatch";
 import {
   LostAndFountNotification,
   MissingStatus,
@@ -20,24 +15,11 @@ import { useDeepCompareEffect } from "~/hooks/useDeepCompareEffect";
 import { AppRoutePaths } from "~/routes/AppRoutePaths";
 import { requireAuthToken } from "~/util/authUtil";
 import { invariantResponse } from "~/util/invariant";
-import {
-  CONFIRM_RENEW_PARAM_KEY,
-  ITEM_PARAM_KEY,
-} from "~/util/searchParamsKeys";
-import {
-  CA_MembershipStatus,
-  MembershipStatus,
-} from "./types/PetServicesTypes";
+
 import { useLostAndFoundReportViewModel } from "./useLostAndFoundReportViewModel";
 import { PET_DOCUMENT_TYPES_LIST } from "./utils/petDocumentConstants";
-import { PET_WATCH_OFFERS, PET_WATCH_TAGS } from "./utils/petServiceConstants";
-import { PET_WATCH_SERVICES_DETAILS } from "./utils/petServiceDetails";
-import { getStatus } from "./utils/petServiceStatusUtils";
-import {
-  PET_WATCH_ANNUAL_UNAVAILABLE_OPTIONS,
-  PetWatchOptionBasedOnMembershipStatus_CA,
-  PetWatchOptionBasedOnMembershipStatus_US,
-} from "./utils/petWatchConstants";
+
+import { useRenewMembershipViewModel } from "./useRenewMembershipViewModel";
 
 export const loader = (({ params }) => {
   const { petId } = params;
@@ -66,7 +48,6 @@ export const loader = (({ params }) => {
 
 export const usePetProfileLayoutViewModel = () => {
   const navigate = useNavigate();
-  const [searchParams, setSearchParams] = useSearchParams();
   const {
     documentTypes,
     lostAndFoundNotificationsPromise,
@@ -104,6 +85,11 @@ export const usePetProfileLayoutViewModel = () => {
     };
   })();
 
+  const renewMembershipViewModel = useRenewMembershipViewModel({
+    selectedPet,
+    postRenew,
+  });
+
   const fetchLostPetHistory = useCallback(async () => {
     setIsLoadingLostPetHistory(true);
 
@@ -138,223 +124,14 @@ export const usePetProfileLayoutViewModel = () => {
     navigate(AppRoutePaths.petEdit);
   };
 
-  const getSelectedPetLocale = () => {
-    const locale = petInfo?.locale ?? "US"; //
-    return { locale };
-  };
-
-  const getPetServiceInfo = () => {
-    const { locale } = getSelectedPetLocale();
-
-    const membershipStatus = selectedPet?.membershipStatus;
-
-    // Safe guard to avoid inconsistencies from the API bringing annual membership for Canadian animals
-    if (
-      membershipStatus?.toLocaleLowerCase().includes("annual") &&
-      locale === "CA"
-    )
-      return { membershipStatus: null, serviceStatus: null };
-
-    const serviceStatus = getStatus({
-      products: selectedPet?.products,
-      locale,
-      membershipStatus,
-    });
-
-    return { membershipStatus, serviceStatus };
-  };
-
-  const getPetWatchInfo = () => {
-    const { membershipStatus, serviceStatus } = getPetServiceInfo();
-
-    if (serviceStatus === null) {
-      return { petWatchOffersAndTags: null, membershipStatus };
-    }
-
-    const petWatchOffersAndTags = {
-      ...PET_WATCH_OFFERS[serviceStatus],
-      ...PET_WATCH_TAGS[serviceStatus],
-    };
-
-    return { petWatchOffersAndTags, membershipStatus };
-  };
-
-  const getPetWatchBenefits = () => {
-    const { locale } = getSelectedPetLocale();
-
-    const petWatchInfo = getPetServiceInfo();
-    const membershipStatus = petWatchInfo?.membershipStatus ?? "Not a member";
-
-    const PetWatchOptionsBasedOnLocale: Record<string, PetCardPetWatchProps[]> =
-      {
-        US: PetWatchOptionBasedOnMembershipStatus_US[
-          membershipStatus as MembershipStatus
-        ],
-        CA: PetWatchOptionBasedOnMembershipStatus_CA[
-          membershipStatus as CA_MembershipStatus
-        ],
-      };
-
-    const petWatchAvailableBenefits = getAvailableBenefitsBasedOnStatus(
-      PetWatchOptionsBasedOnLocale[locale]
-    );
-
-    const petWatchAnnualUnavailableBenefits =
-      getUnavailableBenefitsBasedOnStatus(petWatchAvailableBenefits);
-
-    return { petWatchAvailableBenefits, petWatchAnnualUnavailableBenefits };
-  };
-
-  const getAvailableBenefitsBasedOnStatus = (
-    petWatchAvailableBenefits: PetCardPetWatchProps[]
-  ) => {
-    const membershipStatus = selectedPet?.membershipStatus?.toLowerCase();
-    const products = selectedPet?.products;
-
-    const hasAnnualProduct = products?.some((product) => {
-      const id = product.id.toLowerCase();
-      // Fragile implementation, should be refactored to get this from the server
-      return id.includes("dogs") || id.includes("cats");
-    });
-
-    // If not a member, return all benefits as is
-    if (membershipStatus?.includes("not a member")) {
-      return petWatchAvailableBenefits;
-    }
-
-    // If annual member with no products, return only direct-connect and recovery-specialists
-    if (membershipStatus?.includes("annual")) {
-      if (!products?.length || hasAnnualProduct)
-        return petWatchAvailableBenefits.filter(
-          (benefit) =>
-            benefit.id === "direct-connect" ||
-            benefit.id === "recovery-specialists"
-        );
-    }
-
-    // If there are products, check for expired ones and update benefits accordingly
-    if (products && products.length > 0) {
-      const expiredProductIds = new Set(
-        products
-          .filter(
-            (product): product is NonNullable<typeof product> =>
-              !!product && !!product.id && product.isExpired
-          )
-          .map((product) => product.id)
-      );
-
-      return petWatchAvailableBenefits.map((benefit) => {
-        if (benefit.id && expiredProductIds.has(benefit.id)) {
-          return { ...benefit, isExpired: true };
-        }
-        return benefit;
-      });
-    }
-
-    // Default case
-    return petWatchAvailableBenefits;
-  };
-
-  const getUnavailableBenefitsBasedOnStatus = (
-    petWatchAvailableBenefits: PetCardPetWatchProps[]
-  ) => {
-    const membershipStatus = selectedPet?.membershipStatus?.toLowerCase();
-
-    if (membershipStatus?.includes("annual")) {
-      const availableBenefitIds = new Set(
-        petWatchAvailableBenefits.map((benefit) => benefit.id)
-      );
-
-      return PET_WATCH_ANNUAL_UNAVAILABLE_OPTIONS.filter(
-        (option) => !availableBenefitIds.has(option.id)
-      );
-    }
-  };
-
-  const getContentDetails = (benefits: PetCardPetWatchProps[]) => {
-    const selectedContent = searchParams.get(ITEM_PARAM_KEY);
-    let contentDetails = PET_WATCH_SERVICES_DETAILS.find(
-      ({ id }) => id === selectedContent
-    );
-
-    const selectedBenefit = benefits.find(
-      (benefit) => benefit.id === selectedContent
-    );
-    if (selectedBenefit && selectedBenefit.isExpired && contentDetails) {
-      contentDetails = {
-        ...contentDetails,
-        primaryAction: {
-          buttonLabel: "Renew service",
-          confirmButtonLabel: "Confirm Renew",
-          message: `Would you like to renew ${contentDetails.title} for another year? This ensures continued protection for your pet's.`,
-          title: "Confirm Renewal",
-          onClick: () => {
-            openModal("renew");
-          },
-        },
-      };
-    }
-
-    return contentDetails;
-  };
-
-  const handleContentChange = (label?: string) => {
-    return () => {
-      if (label) {
-        setSelectedContent(label);
-      } else {
-        resetContent();
-      }
-    };
-  };
-
-  const setSelectedContent = (value: string) => {
-    searchParams.set(ITEM_PARAM_KEY, value);
-    setSearchParams(searchParams);
-  };
-
-  const resetContent = () => {
-    searchParams.delete(ITEM_PARAM_KEY);
-    setSearchParams(searchParams);
-  };
-
-  const openModal = (modalType: string) => {
-    searchParams.set(CONFIRM_RENEW_PARAM_KEY, modalType);
-    setSearchParams(searchParams);
-  };
-
-  const closeConfirmRenewModal = () => {
-    searchParams.delete(CONFIRM_RENEW_PARAM_KEY);
-    setSearchParams(searchParams);
-  };
-
-  const isConfirmRenewModalOpen = searchParams.has(CONFIRM_RENEW_PARAM_KEY);
-
-  const onRenewMembership = async () => {
-    const selectedContent = searchParams.get(ITEM_PARAM_KEY);
-    if (!selectedContent) return null;
-
-    return await postRenew({
-      autoRenew: true,
-      id: selectedContent,
-      petId,
-    });
-  };
-
   return {
-    closeConfirmRenewModal,
     documentTypes,
-    getContentDetails,
-    handleContentChange,
-    isConfirmRenewModalOpen,
     isLoading,
     lostPetHistory,
     onEditPet,
-    onRenewMembership,
     pet: selectedPet,
-    petWatchBenefits: getPetWatchBenefits(),
-    petWatchInfo: getPetWatchInfo(),
     ...lostAndFoundViewModel,
+    ...renewMembershipViewModel,
   };
 };
